Cost Scaling based on generation
Value can be controlled on the Library&Rules object

The implementation was made low level as an overload for AdvanceStat of BonsaiVar Struct

+++ Assets/Script/Data&Rules/MathfExt.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public static class MathfExt
 {
-    public static float BonsaiFunct(float baseVal, float scale, int lvl)
+    public static float BonsaiFunct(float baseVal, float scale, float lvl)
     {
         return baseVal * Mathf.Pow(scale, (lvl - 1));       
     }
@@ -12,9 +12,8 @@ public static class MathfExt
     /// <summary>
     /// Checks if a float value fits between the range between positive and negative of another float value.
     /// </summary>
-    /// <param name="a"> tested float </param>
-    /// <param name="signlessLimit"> float marking range of testing (from -signlessLimit to signlessLimit) </param>
-    /// <returns> true - x within range; false - x out of range </returns>
+    /// <param name="a">tested float</param>
+    /// <param name="signlessLimit">float marking range of testing (from -signlessLimit to signlessLimit)</param>
     public static bool IsBoundBy(float a, float signlessLimit)
     {
         return Mathf.Abs(a) < signlessLimit;
@@ -26,7 +25,6 @@ public static class MathfExt
     /// <param name="x"> tested float </param>
     /// <param name="min"> float marking the lower bound of range of testing </param>
     /// <param name="max"> float marking the higher bound of range of testing </param>
-    /// <returns> true - x within range; false - x out of range </returns>
     public static bool IsBoundBy(float x, float min, float max)
     {
         return Mathf.Abs(x-(max+min)/2) < (max-min)/2;
